+++ Assets/Scripts/SS3D/Systems/Entities/PlayerControllable.cs
@@ -1,5 +1,4 @@
 ﻿using FishNet.Object;
-using SS3D.Systems.Inventory;
 using UnityEngine;
 
 namespace SS3D.Systems.Entities
@@ -7,11 +6,8 @@ namespace SS3D.Systems.Entities
     /// <summary>
     /// Base class for all things that can be controlled by a player
     /// </summary>
-    public class Entity : NetworkBehaviour
+    public class PlayerControllable : NetworkBehaviour
     {
-        [SerializeField] private EntityInventory _inventory;
-
-        public EntityInventory Inventory => _inventory;
         
     }
 }
