--- Assets/Scripts/SS3D/Systems/Inventory/InventorySlot.cs
@@ -1,9 +1,0 @@
-﻿using UnityEngine;
-
-namespace SS3D.Systems.Inventory
-{
-    public class InventorySlot : MonoBehaviour
-    {
-        
-    }
-}
