+++ Assets/Scripts/SS3D/Systems/Storage/EntityInventory.cs
@@ -2,14 +2,14 @@
 using System.Linq;
 using UnityEngine;
 
-namespace SS3D.Systems.Inventory
+namespace SS3D.Systems.Storage
 {
     /// <summary>
     /// Base class for all entities's inventory
     /// </summary>
     public class EntityInventory : MonoBehaviour
     {
-        [SerializeField] private List<InventorySlot> _slots;
+        [SerializeField] private List<Storage.Inventory> _slots;
 
         /// <summary>
         /// Gets clothing slots from inventory, use T as the specific type of slot
