+++ Assets/Scripts/SS3D/Systems/Storage/EntityInventory.cs.meta
