+++ Assets/Scripts/SS3D/Systems/Storage/Inventory.cs
@@ -1,0 +1,9 @@
+﻿using UnityEngine;
+
+namespace SS3D.Systems.Storage
+{
+    public class Inventory : MonoBehaviour
+    {
+        
+    }
+}
