+++ Assets/Scripts/SS3D/Systems/Storage/Inventory.cs.meta
